moved class to seperate file
